Add Australian Synchrotron MX1 configuration (#27)
Checked against actual dataset. Note the addition of an `_diffrn_source.[local]_config_id` data name so that we can track different configurations of the same beamline.
Also, two theta was at zero for the test data set so the sense of movement of that axis is assumed to be the same as phi.